+++ test/list_test.dart
@@ -361,4 +361,27 @@ Future<void> main([List<String>? args]) async {
     await Future<void>.delayed(Duration(milliseconds: 10));
     realm.close();
   });
+
+  
+  test('List query', () {
+    final config = Configuration([Team.schema, Person.schema]);
+    final realm = Realm(config);
+
+    final person = Person('John');
+    final team = Team('team1', players: [
+      Person('Pavel'),
+      person,
+      Person('Alex'),
+    ]);
+
+    realm.write(() => realm.add(team));
+
+    // TODO: Get rid of cast, once type signature of team.players is a RealmList<Person>
+    // as opposed to the List<Person> we have today.
+    final result = (team.players as RealmList<Person>).query(r'name BEGINSWITH $0', ['J']);
+
+    expect(result, [person]);
+
+    realm.close();
+  });
 }
